# src/async_api_client/openapi/openapi-3.0.3.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Airflow API (Stable)",
...
    "version": "2.10.5",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "contact": {
      "name": "Apache Software Foundation",
      "url": "https://airflow.apache.org",
      "email": "dev@airflow.apache.org"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Apache Airflow Stable API."
    }
  ],
  "paths": {
...
    "/dags": {
      "get": {
        "summary": "List DAGs",
        "description": "List DAGs in the database.\n`dag_id_pattern` can be set to match dags of a specific pattern\n",
        "x-openapi-router-controller": "airflow.api_connexion.endpoints.dag_endpoint",
        "operationId": "get_dags",
        "tags": [
          "DAG"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PageLimit"
          },
          {
            "$ref": "#/components/parameters/PageOffset"
          },
          {
            "$ref": "#/components/parameters/OrderBy"
          },
          {
            "$ref": "#/components/parameters/FilterTags"
          },
          {
            "$ref": "#/components/parameters/OnlyActive"
          },
          {
            "$ref": "#/components/parameters/Paused"
          },
          {
            "$ref": "#/components/parameters/ReturnFields"
          },
          {
            "name": "dag_id_pattern",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "If set, only return DAGs with dag_ids matching this pattern.\n"
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DAGCollection"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          }
        }
      },
      "patch": {
        "summary": "Update DAGs",
        "description": "Update DAGs of a given dag_id_pattern using UpdateMask.\nThis endpoint allows specifying `~` as the dag_id_pattern to update all DAGs.\n*New in version 2.3.0*\n",
        "x-openapi-router-controller": "airflow.api_connexion.endpoints.dag_endpoint",
        "operationId": "patch_dags",
        "parameters": [
          {
            "$ref": "#/components/parameters/PageLimit"
          },
          {
            "$ref": "#/components/parameters/PageOffset"
          },
          {
            "$ref": "#/components/parameters/FilterTags"
          },
          {
            "$ref": "#/components/parameters/UpdateMask"
          },
          {
            "$ref": "#/components/parameters/OnlyActive"
          },
          {
            "name": "dag_id_pattern",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "If set, only update DAGs with dag_ids matching this pattern.\n"
          }
        ],
        "tags": [
          "DAG"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DAG"
              },
              "example": {
                "is_paused": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DAGCollection"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/dags/{dag_id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/DAGID"
        }
      ],
      "get": {
        "summary": "Get basic information about a DAG",
        "description": "Presents only information available in database (DAGModel).\nIf you need detailed information, consider using GET /dags/{dag_id}/details.\n",
        "x-openapi-router-controller": "airflow.api_connexion.endpoints.dag_endpoint",
        "operationId": "get_dag",
        "tags": [
          "DAG"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ReturnFields"
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DAG"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "summary": "Update a DAG",
        "x-openapi-router-controller": "airflow.api_connexion.endpoints.dag_endpoint",
        "operationId": "patch_dag",
        "parameters": [
          {
            "$ref": "#/components/parameters/UpdateMask"
          }
        ],
        "tags": [
          "DAG"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DAG"
              },
              "example": {
                "is_paused": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DAG"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "summary": "Delete a DAG",
        "description": "Deletes all metadata related to the DAG, including finished DAG Runs and Tasks.\nLogs are not deleted. This action cannot be undone.\n\n*New in version 2.2.0*\n",
        "x-openapi-router-controller": "airflow.api_connexion.endpoints.dag_endpoint",
        "operationId": "delete_dag",
        "tags": [
          "DAG"
        ],
        "responses": {
          "204": {
            "description": "Success."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/AlreadyExists"
          }
        }
      }
    },
...
    "/dags/{dag_id}/dagRuns": {
      "parameters": [
        {
          "$ref": "#/components/parameters/DAGID"
        }
      ],
      "get": {
        "summary": "List DAG runs",
        "description": "This endpoint allows specifying `~` as the dag_id to retrieve DAG runs for all DAGs.\n",
        "x-openapi-router-controller": "airflow.api_connexion.endpoints.dag_run_endpoint",
        "operationId": "get_dag_runs",
        "tags": [
          "DAGRun"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PageLimit"
          },
          {
            "$ref": "#/components/parameters/PageOffset"
          },
          {
            "$ref": "#/components/parameters/FilterExecutionDateGTE"
          },
          {
            "$ref": "#/components/parameters/FilterExecutionDateLTE"
          },
          {
            "$ref": "#/components/parameters/FilterStartDateGTE"
          },
          {
            "$ref": "#/components/parameters/FilterStartDateLTE"
          },
          {
            "$ref": "#/components/parameters/FilterEndDateGTE"
          },
          {
            "$ref": "#/components/parameters/FilterEndDateLTE"
          },
          {
            "$ref": "#/components/parameters/FilterUpdatedAtGTE"
          },
          {
            "$ref": "#/components/parameters/FilterUpdatedAtLTE"
          },
          {
            "$ref": "#/components/parameters/FilterState"
          },
          {
            "$ref": "#/components/parameters/OrderBy"
          },
          {
            "$ref": "#/components/parameters/ReturnFields"
          }
        ],
        "responses": {
          "200": {
            "description": "List of DAG runs.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DAGRunCollection"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          }
        }
      },
      "post": {
        "summary": "Trigger a new DAG run.",
        "description": "This will initiate a dagrun. If DAG is paused then dagrun state will remain queued, and the task won't run.\n",
        "x-openapi-router-controller": "airflow.api_connexion.endpoints.dag_run_endpoint",
        "operationId": "post_dag_run",
        "tags": [
          "DAGRun"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DAGRun"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DAGRun"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "409": {
            "$ref": "#/components/responses/AlreadyExists"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/dags/{dag_id}/dagRuns/{dag_run_id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/DAGID"
        },
        {
          "$ref": "#/components/parameters/DAGRunID"
        }
      ],
      "get": {
        "summary": "Get a DAG run",
        "x-openapi-router-controller": "airflow.api_connexion.endpoints.dag_run_endpoint",
        "operationId": "get_dag_run",
        "tags": [
          "DAGRun"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ReturnFields"
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DAGRun"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "summary": "Delete a DAG run",
        "x-openapi-router-controller": "airflow.api_connexion.endpoints.dag_run_endpoint",
        "operationId": "delete_dag_run",
        "tags": [
          "DAGRun"
        ],
        "responses": {
          "204": {
            "description": "Success."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "summary": "Modify a DAG run",
        "description": "Modify a DAG run.\n\n*New in version 2.2.0*\n",
        "x-openapi-router-controller": "airflow.api_connexion.endpoints.dag_run_endpoint",
        "operationId": "update_dag_run_state",
        "tags": [
          "DAGRun"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDagRunState"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DAGRun"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/dags/{dag_id}/dagRuns/{dag_run_id}/upstreamDatasetEvents": {
      "parameters": [
        {
          "$ref": "#/components/parameters/DAGID"
        },
        {
          "$ref": "#/components/parameters/DAGRunID"
        }
      ],
      "get": {
        "summary": "Get dataset events for a DAG run",
        "description": "Get datasets for a dag run.\n\n*New in version 2.4.0*\n",
        "x-openapi-router-controller": "airflow.api_connexion.endpoints.dag_run_endpoint",
        "operationId": "get_upstream_dataset_events",
        "tags": [
          "DAGRun",
          "Dataset"
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetEventCollection"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DAG": {
        "type": "object",
        "description": "DAG",
        "properties": {
          "dag_id": {
            "type": "string",
            "readOnly": true,
            "description": "The ID of the DAG."
          },
          "dag_display_name": {
            "type": "string",
            "readOnly": true,
            "description": "Human centric display text for the DAG.\n\n*New in version 2.9.0*\n"
          },
          "root_dag_id": {
            "type": "string",
            "readOnly": true,
            "nullable": true,
            "description": "If the DAG is SubDAG then it is the top level DAG identifier. Otherwise, null."
          },
          "is_paused": {
            "type": "boolean",
            "nullable": true,
            "description": "Whether the DAG is paused."
          },
          "is_active": {
            "description": "Whether the DAG is currently seen by the scheduler(s).\n\n*New in version 2.1.1*\n\n*Changed in version 2.2.0*&#58; Field is read-only.\n",
            "nullable": true,
            "readOnly": true,
            "type": "boolean"
          },
          "is_subdag": {
            "description": "Whether the DAG is SubDAG.",
            "type": "boolean",
            "readOnly": true
          },
          "last_parsed_time": {
            "type": "string",
            "format": "date-time",
            "readOnly": true,
            "nullable": true,
            "description": "The last time the DAG was parsed.\n\n*New in version 2.3.0*\n"
          },
          "last_pickled": {
            "type": "string",
            "format": "date-time",
            "readOnly": true,
            "nullable": true,
            "description": "The last time the DAG was pickled.\n\n*New in version 2.3.0*\n"
          },
          "last_expired": {
            "type": "string",
            "format": "date-time",
            "readOnly": true,
            "nullable": true,
            "description": "Time when the DAG last received a refresh signal\n(e.g. the DAG's \"refresh\" button was clicked in the web UI)\n\n*New in version 2.3.0*\n"
          },
          "scheduler_lock": {
            "type": "boolean",
            "readOnly": true,
            "nullable": true,
            "description": "Whether (one of) the scheduler is scheduling this DAG at the moment\n\n*New in version 2.3.0*\n"
          },
          "pickle_id": {
            "type": "string",
            "readOnly": true,
            "nullable": true,
            "description": "Foreign key to the latest pickle_id\n\n*New in version 2.3.0*\n"
          },
          "default_view": {
            "type": "string",
            "nullable": true,
            "readOnly": true,
            "description": "Default view of the DAG inside the webserver\n\n*New in version 2.3.0*\n"
          },
          "fileloc": {
            "description": "The absolute path to the file.",
            "type": "string",
            "readOnly": true
          },
          "file_token": {
            "type": "string",
            "readOnly": true,
            "description": "The key containing the encrypted path to the file. Encryption and decryption take place only on the server. This prevents the client from reading an non-DAG file. This also ensures API extensibility, because the format of encrypted data may change.\n"
          },
          "owners": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "readOnly": true
          },
          "description": {
            "type": "string",
            "readOnly": true,
            "nullable": true,
            "description": "User-provided DAG description, which can consist of several sentences or paragraphs that describe DAG contents.\n"
          },
          "schedule_interval": {
            "$ref": "#/components/schemas/ScheduleInterval"
          },
          "timetable_description": {
            "type": "string",
            "readOnly": true,
            "nullable": true,
            "description": "Timetable/Schedule Interval description.\n\n*New in version 2.3.0*\n"
          },
          "tags": {
            "description": "List of tags.",
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Tag"
            },
            "readOnly": true
          },
          "max_active_tasks": {
            "type": "integer",
            "nullable": true,
            "readOnly": true,
            "description": "Maximum number of active tasks that can be run on the DAG\n\n*New in version 2.3.0*\n"
          },
          "max_active_runs": {
            "type": "integer",
            "nullable": true,
            "readOnly": true,
            "description": "Maximum number of active DAG runs for the DAG\n\n*New in version 2.3.0*\n"
          },
          "has_task_concurrency_limits": {
            "type": "boolean",
            "nullable": true,
            "readOnly": true,
            "description": "Whether the DAG has task concurrency limits\n\n*New in version 2.3.0*\n"
          },
          "has_import_errors": {
            "type": "boolean",
            "nullable": true,
            "readOnly": true,
            "description": "Whether the DAG has import errors\n\n*New in version 2.3.0*\n"
          },
          "next_dagrun": {
            "type": "string",
            "format": "date-time",
            "readOnly": true,
            "nullable": true,
            "description": "The logical date of the next dag run.\n\n*New in version 2.3.0*\n"
          },
          "next_dagrun_data_interval_start": {
            "type": "string",
            "format": "date-time",
            "readOnly": true,
            "nullable": true,
            "description": "The start of the interval of the next dag run.\n\n*New in version 2.3.0*\n"
          },
          "next_dagrun_data_interval_end": {
            "type": "string",
            "format": "date-time",
            "readOnly": true,
            "nullable": true,
            "description": "The end of the interval of the next dag run.\n\n*New in version 2.3.0*\n"
          },
          "next_dagrun_create_after": {
            "type": "string",
            "format": "date-time",
            "readOnly": true,
            "nullable": true,
            "description": "Earliest time at which this ``next_dagrun`` can be created.\n\n*New in version 2.3.0*\n"
          },
          "max_consecutive_failed_dag_runs": {
            "type": "integer",
            "nullable": true,
            "readOnly": true,
            "description": "(experimental) The maximum number of consecutive DAG failures before DAG is automatically paused.\n\n*New in version 2.9.0*\n"
          }
        }
      },
      "DAGCollection": {
        "description": "Collection of DAGs.\n\n*Changed in version 2.1.0*&#58; 'total_entries' field is added.\n",
        "type": "object",
        "allOf": [
          {
            "type": "object",
            "properties": {
              "dags": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/DAG"
                }
              }
            }
          },
          {
            "$ref": "#/components/schemas/CollectionInfo"
          }
        ]
      },
...
    },
    "parameters": {
      "PageOffset": {
        "in": "query",
        "name": "offset",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 0
        },
        "description": "The number of items to skip before starting to collect the result set."
      },
      "PageLimit": {
        "in": "query",
        "name": "limit",
        "required": false,
        "schema": {
          "type": "integer",
          "default": 100
        },
        "description": "The numbers of items to return."
      },
...
      "DAGID": {
        "in": "path",
        "name": "dag_id",
        "schema": {
          "type": "string"
        },
        "required": true,
        "description": "The DAG ID."
      },
...
      "FilterExecutionDateGTE": {
        "in": "query",
        "name": "execution_date_gte",
        "schema": {
          "type": "string",
          "format": "date-time"
        },
        "required": false,
        "description": "Returns objects greater or equal to the specified date.\n\nThis can be combined with execution_date_lte parameter to receive only the selected period.\n"
      },
      "FilterExecutionDateLTE": {
        "in": "query",
        "name": "execution_date_lte",
        "schema": {
          "type": "string",
          "format": "date-time"
        },
        "required": false,
        "description": "Returns objects less than or equal to the specified date.\n\nThis can be combined with execution_date_gte parameter to receive only the selected period.\n"
      },
...
      "FilterState": {
        "in": "query",
        "name": "state",
        "schema": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "required": false,
        "description": "The value can be repeated to retrieve multiple matching values (OR condition)."
      },
...
      "FilterTags": {
        "in": "query",
        "name": "tags",
        "schema": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "description": "List of tags to filter results.\n\n*New in version 2.2.0*\n"
      },
...
      "OrderBy": {
        "in": "query",
        "name": "order_by",
        "schema": {
          "type": "string"
        },
        "required": false,
        "description": "The name of the field to order the results by.\nPrefix a field name with `-` to reverse the sort order.\n\n*New in version 2.1.0*\n"
      },
      "OnlyActive": {
        "in": "query",
        "name": "only_active",
        "schema": {
          "type": "boolean",
          "default": true
        },
        "required": false,
        "description": "Only filter active DAGs.\n\n*New in version 2.1.1*\n"
      },
...
      "Paused": {
        "in": "query",
        "name": "paused",
        "schema": {
          "type": "boolean"
        },
        "required": false,
        "description": "Only filter paused/unpaused DAGs. If absent or null, it returns paused and unpaused DAGs.\n\n*New in version 2.6.0*\n"
      },
...
      "UpdateMask": {
        "in": "query",
        "name": "update_mask",
        "schema": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "description": "The fields to update on the resource. If absent or empty, all modifiable fields are updated.\nA comma-separated list of fully qualified names of fields.\n",
        "style": "form",
        "explode": false
      },
      "ReturnFields": {
        "in": "query",
        "name": "fields",
        "schema": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "description": "List of field for return.\n"
      }
    },
...
    "responses": {
      "BadRequest": {
        "description": "Client specified an invalid argument.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthenticated": {
        "description": "Request not authenticated due to missing, invalid, authentication info.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "PermissionDenied": {
        "description": "Client does not have sufficient permission.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "A specified resource is not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "AlreadyExists": {
        "description": "An existing resource conflicts with the request.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "Basic": {
        "type": "http",
        "scheme": "basic"
      },
      "GoogleOpenId": {
        "type": "openIdConnect",
        "openIdConnectUrl": "https://accounts.google.com/.well-known/openid-configuration"
      },
      "Kerberos": {
        "type": "http",
        "scheme": "negotiate"
      }
    }
  },
  "security": [],
...
}